requirements.txt
streamlit>=1.28.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
langchain>=0.2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0
numpy>=1.24.0
ollama>=0.1.8